Redesign transliterators part 1
Now they use std::array instead of std::unordered_map
so they don't require an allocation in the constructor.
So now translit class is pure static.
There is still a possible allocation in the translit buffer
it is changed in the following commits: part 2 and part 3.
This also makes transliterators more than two times faster.