Add overflow protection
The library doesn't use exceptions but this case is special
and must be handled like this.
With this change we simply emulate what C++ standard library
will throw when std::string.resize(std::string.max_size() + 1)
if it would be possible (max_size() + 1 doesn't overflow).
This case is simple impossible on x64 and almost impossible
on x86 (a program will throw bad alloc much sooner anyway).
So with this we only handle the case that might occur in theory
but never occur in practice.