Redesign arrays usage in low-level normalization
This is significant change of the low-level normalization code
that allows to use std::array (or C++ safe array with checks)
instead of C arrays on the low-level.

This change makes functions a bit bigger
but the performance is the same.
For example byte size of impl_norm_to_nfc_utf8 function:
old: 3840 new: 4128 according to GCC nm.