Add safe layer
Initial implementation of the safe layer for the low-level.

The first part of the safe layer is implemented - safe arrays.
The second part is not implemented in this commit - safe iterators.

By default the safe layer uses std::array
but able to use safe::array with bounds checks.

Define UNI_ALGO_FORCE_C_ARRAYS can be used
to force to use C arrays (old behaviour) if needed for tests.
This is enabled in this commit,
disabled in the following commit "Enable safe layer".

Observations:
std::array version produces a bit larger executable
compared to C arrays version.
Around 10 KB larger for test executable (GCC and MSVC tested).
The performance is the same, that's the way it meant to be.