Add safe iterators to safe layer
Initial implementation of safe iterators for the safe layer.

This is the second and most complex part of the safe layer
The first one, safe arrays, is added in commit: c60e1a314b6963305823b8b5a4d66516f2060e49

Use of the safe iterators will be added in following commits.

Note that the performance impact of the safe iterators
is much more significant than the safe arrays.
Around 5-20% worse for different functions and can be
even worse if they will prevent a compiler to vectorize stuff.
So they should not be used in release by default.