Modify templates for reverse iterators in low-level
This is a workaround for the following commits.
Reverse iterator functions should not support sentinel
but it is added in this commit to fix compilation.

The problem is case_final_sigma_utf8/16 function in impl_case.h
that handles both forward and reverse algorithms
and doesn't compile with sentinel without this change.

In the future it is better to split this function
to forward and reverse version and revert this commit.