Use pass by const reference
Even though those template functions always used with
std::string_view use pass by const reference
because it can be a problem in the future as it always happens
for example if it will be used with std::string for some reason.
It's not like const reference there will ever affect performance.

Also similar template functions in conv.h are not affected
by this change because they already use pass by const reference.