Add likely for better performance
The reason of this change because I noticed strange performance drop
in some cases in GCC 12.1.0 compared to old result with GCC 7.3.0

I'm not sure what is going on but likely on that 0xE0 branch
gives the better performance in all cases in GCC/Clang.
More noticeable in GCC and 3 bytes UTF-8 sequences. Almost 10% better.
It seems likely there just reorders branches better for UTF-8.
Usually it's better not to mess with likely like this
but this time it's better for literally all cases so it should be fine.

UTF-8 conversion and normalization tested.

GCC 12.1.0 x64 (very noticeable)
Clang 14.0.6 x64 (noticeable)
GCC 7.3.0 x64 (almost not noticeable)

ASM GCC 12.1.0 x64 -O3 observations:
Without likely impl_utf8to16 function
uses 1 extra register ASM: push %r13
it should not be a reason for the better performance
with likely just mentioning it here.
I didn't notice anything unusual except that.
And yes it does reorder the branches as expected.