Introduce comment tags in low-level
Now we require that all variables must be initialized but
it makes things less readable the problem is 0 and false.
int something = 0;
bool something = false;
Do they initialized just for the sake of initialization
or an algorithm requires them to be initialized with this value.
It is very important in the low-level because many Unicode algorithms
are very complex and it makes it harder to understand them.
So this commit introduces tags in comments for the low-level
for this case and some other cases.

All tags is added in the following commits.

For reference this is the list of commits where zero initialization was
added for variables/arrays that can be uninitialized in the low-level:
83d51a566c03ba9fd67dfdcbf3c3dc8066c2de4a
9e683d9ff0e86acded7603cbe65395ceb046817c
1fece5bead60d4976d8d7f953b4f6afa184bee04
This was already done for some of variables even before
initial commit so I will tag them from my memory.