Make tests constexpr part 1
This makes all test functions return bool (part 1)
and TESTX macro return false if a test fails (part 2)
to make it possible to test in constexpr context
without many modifications of tests.
A bit ugly but it's the easiest way to achieve it.

Some more changes needed for tests to make it work
it is done in the following commit
because this commit is already big enough.