Update README.md
Change the alphabet in the collation function example.

The Macedonian alphabet is not good for the example because
it's locale-dependent and the result is incorrect
for our simple locale-independent collation function.
So use the Ukrainian alphabet instead that is locale-independent
and contains many letters that is out of order with binary comparison
so it's good for the demonstration.