Add template parameter for ASCII optimization
This adds ability to disable ASCII optimization
with the extra template parameter to restore ability
of conversion functions to work with any types of iterators.

See example in the next commit.

Note that extra template parameter added for all low-level
conversion functions but not all of them support ASCII optimization.
This is to make them easier to use as generic functions in a wrapper.