Add typename keyword for Clang 14
Continuation of the previous commit.
Add extra typename keyword in one more place where
resize_and_overwrite is used, it seems Clang 14 needs it there.

Also this is the first commit signed with a different GPG key because
I forgot passphrase to the old one, but it's still me I swear :)